# openapi.yml
openapi: 3.0.0
info:
  description: a blog api
  version: "1.0.0"
  title: blog
  contact:
    email: young@young-zy.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/young-zy/blog/1.0.0
...
paths:
...
  /question:
    post:
      tags:
        - question
      summary: 'add a new question'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: won't be shown, used only for notification when answered
                  type: string
                  format: email
                questionContent:
                  type: string
      responses:
        200:
          description: question added successfully
    get:
      summary: get the question list
      tags:
        - question
      parameters:
        - name: page
          in: query
          description: page of the reply, default value is 1
          schema:
            type: integer
            format: int32
        - name: amount
          in: query
          description: amount of the results per page, default value is 10
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
components:
  schemas:
...
    Question:
      type: object
      required:
        - questionContent
      properties:
        questionContent:
          type: string
        replyContent:
          type: string
...